+++ Models/AdminVM/DashboardViewModel.cs
@@ -11,7 +11,7 @@ public class Widgets
 {
     public int? UsersCount {get;set;} = 0;
     public int? SalesCount {get;set;} = 0;
-    public int? Revenue {get;set;} = 0;
+    public decimal? Revenue {get;set;} = 0;
     public int? TotalProductQuantity {get;set;} = 0;  
 }
